refactor: extract list nodes
